Change weight type to decimal

+++ employee/models.py
@@ -105,7 +105,7 @@ class Employee(AbstractUser):
 class Parameter(BaseModel):
     question = models.CharField(max_length=100)
     department = models.ForeignKey(Department, on_delete=models.CASCADE, related_name='departments')
-    weight = models.IntegerField(default=0)
+    weight = models.DecimalField(default=0, decimal_places=2, max_digits=12)
 
     def __str__(self):
         return self.question
